Fix Go 1.25 compatibility: cast stat.Type from uint32 to int64 for map index

+++ mount_predicate.go
@@ -243,7 +243,7 @@ func getFileSystemType(paths ...string) []string {
 			continue
 		}
 
-		fsTypes = append(fsTypes, FsTypeMap[stat.Type])
+		fsTypes = append(fsTypes, FsTypeMap[int64(stat.Type)])
 	}
 
 	return fsTypes
